Create the specialized interface
This expands the refactored script to generate a specialized interface to access
CUTEst. This is specialized in the sense that is easier to use than the core
functions, but retain the naming and parameter position (when present) from
CUTEst.

For each CUTEst function there are two kinds of functions implemented: a
function that returns the output parameters, and one that accepts them as
arguments, modifying them inplace.
When at least one of the inplace arguments is an array of integers, two inplace
functions are created: one with an array of Int's and one with an array of
Cint's. The latter is faster, but the former is the default for 64 bits Julia.

For each of these functions, the problem parameters can be passed individually
or through the CUTEstModel, simplifying the syntax.