Create tests for the specialized interface
Create a script to generated theses tests from `specialized_interface.jl`.
The script needed a lot of "translation" due to some variable names being
used more than once for slightly different meanings.

Note that some function options were not tested, specially for functions
with boolean flags for the transpose of the Jacobian, or to select between
using the objective function or the Lagrangian function.