Fix MANWIDTH on the script and remove hyphens and justification
Inside the script, generating the interface, we need to set MANWIDTH so that
different terminals create the same documentation.
We set the MANWIDTH to a very large value so that hyphenation and justification
between words are also removed because they made the documentation look bad.

Note: We tried using `--nh` and `--nj` but the "less than" symbols were being
not being processed, so this was the workaround.